requirements.txt
albucore==0.0.24
albumentations==2.0.8
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.2.3
decorator==4.4.2
easydict==1.13
eval_type_backport==0.3.1
exceptiongroup==1.3.1
fastapi==0.126.0
ffmpeg==1.4
filelock==3.19.1
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
ftfy==6.3.1
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httplib2==0.31.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.3
matplotlib==3.9.4
ml_dtypes==0.5.4
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnx==1.19.1
onnxruntime==1.19.2
openai-whisper==20240930
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
prettytable==3.16.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.1
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scenedetect==0.6.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
simsimd==6.5.8
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
starlette==0.49.3
stringzilla==4.5.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.1
uvicorn==0.39.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
zipp==3.23.0
